Decompose system/file.go for POSIX and Windows to make Goss buildable on Windows
